Loan are being created(Perfectly)

+++ engine/src/main/java/engine/Entity/LoanGiver.java
@@ -2,6 +2,7 @@ package engine.Entity;
 
 import engine.Preloan;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.concurrent.ExecutionException;
 import org.slf4j.Logger;
@@ -112,11 +113,11 @@ public class LoanGiver extends BlockchainCommunicator{
     }
     
     private BigInteger getInterestScaled(String interest){
-        return (new BigInteger(interest)).multiply(new BigInteger(scale));
+        return (new BigDecimal(interest)).multiply(new BigDecimal(scale)).toBigInteger();
     }
     
     private BigInteger getInterestReciprocal(String interest){
-        return (BigInteger.ONE).divide(new BigInteger(interest));
+        return (BigDecimal.ONE).divide(new BigDecimal(interest)).toBigInteger();
     }
         
     private void informLedger(String contractAddress, String ledgerAddress) throws Exception {
